# server/go/api_wager.go
/*
 * Wager service APIs
 *
 * APIs for a wager system
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package wager

import (
	"net/http"

	"github.com/gin-gonic/gin"
)

// BuyWagersPost - 
func BuyWagersPost(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// CreateWagerPost - 
func CreateWagerPost(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}

// ListWagersGet - 
func ListWagersGet(c *gin.Context) {
	c.JSON(http.StatusOK, gin.H{})
}