# server/go/model_buy_wage_request.go
/*
 * Wager service APIs
 *
 * APIs for a wager system
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package wager

type BuyWageRequest struct {

	BuyingPrice float32 `json:"buying_price,omitempty"`
}