# server/go/model_wager.go
/*
 * Wager service APIs
 *
 * APIs for a wager system
 *
 * API version: 1.0.0
 * Generated by: OpenAPI Generator (https://openapi-generator.tech)
 */

package wager

type Wager struct {

	Id int64 `json:"id,omitempty"`

	TotalWagerValue int32 `json:"total_wager_value,omitempty"`

	Odds int32 `json:"odds,omitempty"`

	SellingPercentage int32 `json:"selling_percentage,omitempty"`

	SellingPrice float32 `json:"selling_price,omitempty"`

	CurrentSellingPrice float32 `json:"current_selling_price,omitempty"`

	PercentageSold int32 `json:"percentage_sold,omitempty"`

	AmountSold float64 `json:"amount_sold,omitempty"`

	PlacedAt int64 `json:"placed_at,omitempty"`
}